Move target net update
