+++ rl_blox/algorithm/ddpg.py
@@ -12,6 +12,7 @@ from flax import nnx
 
 from ..blox.function_approximator.mlp import MLP
 from ..blox.function_approximator.policy_head import DeterministicTanhPolicy
+from ..blox.target_net import soft_target_net_update
 from ..logging.logger import LoggerBase
 
 
@@ -284,42 +285,6 @@ def ddpg_update_actor(
     )(q, observations, policy)
     policy_optimizer.update(grads)
     return actor_loss_value
-
-
-@nnx.jit
-def soft_target_net_update(
-    net: nnx.Module, target_net: nnx.Module, tau: float
-) -> None:
-    r"""Inplace soft update for target network with Polyak averaging.
-
-    The soft update for the target network is supposed to be applied with about
-    the same frequency as the update for the live network.
-
-    Update formula:
-
-    .. math::
-
-        \theta' \leftarrow \tau \theta + (1 - \tau) \theta'
-
-    where :math:`\theta` are the weights of the live network and
-    :math:`\theta'` are the weights of the target network.
-
-    Parameters
-    ----------
-    net : nnx.Module
-        Live network with weights :math:`\theta`.
-
-    target_net : nnx.Module
-        Target network with weights :math:`\theta'`.
-
-    tau : float
-        The step size :math:`\tau`, i.e., the coefficient with which the live
-        network's parameters will be multiplied.
-    """
-    params = nnx.state(net)
-    target_params = nnx.state(target_net)
-    target_params = optax.incremental_update(params, target_params, tau)
-    nnx.update(target_net, target_params)
 
 
 def sample_actions(
