+++ rl_blox/algorithm/sac.py
@@ -14,8 +14,9 @@ from ..blox.function_approximator.policy_head import (
     GaussianTanhPolicy,
     StochasticPolicyBase,
 )
+from ..blox.target_net import soft_target_net_update
 from ..logging.logger import LoggerBase
-from .ddpg import ReplayBuffer, mse_action_value_loss, soft_target_net_update
+from .ddpg import ReplayBuffer, mse_action_value_loss
 
 
 def sac_actor_loss(
